requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python-headless==4.10.0.82
packaging==24.1
pillow==10.3.0
Werkzeug==3.0.3
